+++ src/Block/Element/BlockQuote.php
@@ -52,8 +52,8 @@ class BlockQuote extends AbstractBlock
 
     public function matchesNextLine(Cursor $cursor)
     {
-        if (!$cursor->isIndented() && $cursor->getFirstNonSpaceCharacter() === '>') {
-            $cursor->advanceToFirstNonSpace();
+        if (!$cursor->isIndented() && $cursor->getNextNonSpaceCharacter() === '>') {
+            $cursor->advanceToNextNonSpaceOrTab();
             $cursor->advance();
             $cursor->advanceBySpaceOrTab();