+++ src/CursorState.php
@@ -36,7 +36,7 @@ class CursorState
     /**
      * @var int|null
      */
-    private $firstNonSpaceCache;
+    private $nextNonSpaceCache;
 
     /**
      * @var int
@@ -58,18 +58,18 @@ class CursorState
      * @param int      $length
      * @param int      $currentPosition
      * @param int      $previousPosition
-     * @param int|null $firstNonSpaceCache
+     * @param int|null $nextNonSpaceCache
      * @param int      $indent
      * @param int      $column
      * @param bool     $partiallyConsumedTab
      */
-    public function __construct($line, $length, $currentPosition, $previousPosition, $firstNonSpaceCache, $indent, $column, $partiallyConsumedTab)
+    public function __construct($line, $length, $currentPosition, $previousPosition, $nextNonSpaceCache, $indent, $column, $partiallyConsumedTab)
     {
         $this->line = $line;
         $this->length = $length;
         $this->currentPosition = $currentPosition;
         $this->previousPosition = $previousPosition;
-        $this->firstNonSpaceCache = $firstNonSpaceCache;
+        $this->nextNonSpaceCache = $nextNonSpaceCache;
         $this->indent = $indent;
         $this->column = $column;
         $this->partiallyConsumedTab = $partiallyConsumedTab;
@@ -109,10 +109,22 @@ class CursorState
 
     /**
      * @return int|null
+     *
+     * @deprecated Use getNextNonSpaceCache() instead
      */
     public function getFirstNonSpaceCache()
     {
-        return $this->firstNonSpaceCache;
+        @trigger_error('CursorState::getFirstNonSpaceCache() will be removed in a future 0.x release.  Use getNextNonSpaceCache() instead. See https://github.com/thephpleague/commonmark/issues/280', E_USER_DEPRECATED);
+
+        return $this->nextNonSpaceCache;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getNextNonSpaceCache()
+    {
+        return $this->nextNonSpaceCache;
     }
 
     /**